+++ packages/generator/src/generator-options.ts
@@ -13,7 +13,7 @@ export interface GeneratorOptions {
   serviceMapping?: PathLike;
   useSwagger: boolean;
   writeReadme: boolean;
-  changelogFile?: PathLike;
+  additionalFiles?: string;
   forceOverwrite: boolean;
   clearOutputDir: boolean;
   aggregatorNpmPackageName?: string;
@@ -77,9 +77,9 @@ export const generatorOptionsCli: KeysToOptions = {
     default: false,
     hidden: true
   },
-  changelogFile: {
+  additionalFiles: {
     describe:
-      'Path to file that will be copied into the generated packages under the filename CHANGELOG.md.',
+      'Glob describing additional files to be added to the each generated service directory.',
     type: 'string',
     coerce: resolve,
     normalize: true,