+++ CHANGELOG.md
@@ -14,11 +14,13 @@
 
 ## Compatibility Notes
 
--
+- _OData client generator_: Remove the hidden generator option `changelogFile`.
+Use the new more general options `additionalFiles` instead.
 
 ## New Functionality
 
--
+- Add the option `--additionalFiles <GlobToFiles>` to the generator.
+If specified, all files matching the glob will be copied to each generated service directory.    
 
 ## Improvements